requirements.txt
discord.py==2.3.2
beautifulsoup4==4.12.3
requests==2.31.0
APScheduler==3.10.4
